Move class to new file to try to work around bizarre bug
Why would this fail to compile in production but not when developing?
