+++ SpiNNaker-allocserv/src/main/java/uk/ac/manchester/spinnaker/alloc/SecurityConfig.java
@@ -30,8 +30,7 @@ import static javax.ws.rs.core.Response.status;
 import static javax.ws.rs.core.Response.Status.FORBIDDEN;
 import static org.slf4j.LoggerFactory.getLogger;
 import static org.springframework.http.HttpStatus.UNAUTHORIZED;
-import static uk.ac.manchester.spinnaker.alloc.LocalAuthProviderImpl.isUnsupportedAuthTokenClass;
-import static uk.ac.manchester.spinnaker.alloc.SecurityConfig.AppAuthTransformationFilter.clearToken;
+import static uk.ac.manchester.spinnaker.alloc.AppAuthTransformationFilter.clearToken;
 import static uk.ac.manchester.spinnaker.alloc.SecurityConfig.TrustLevel.USER;
 import static uk.ac.manchester.spinnaker.alloc.SecurityConfigUtils.installInjectableTrustStoreAsDefault;
 import static uk.ac.manchester.spinnaker.alloc.SecurityConfigUtils.loadTrustStore;
@@ -66,7 +65,6 @@ import javax.net.ssl.SSLContext;
 import javax.net.ssl.TrustManager;
 import javax.net.ssl.TrustManagerFactory;
 import javax.net.ssl.X509TrustManager;
-import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -109,7 +107,6 @@ import org.springframework.security.web.authentication.logout.LogoutHandler;
 import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
 import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
 import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
 import org.springframework.web.servlet.ViewResolver;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
@@ -397,91 +394,6 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 				.addLogoutHandler((req, resp, auth) -> clearToken(req))
 				.deleteCookies(SESSION_COOKIE).invalidateHttpSession(true)
 				.logoutSuccessUrl(loginUrl);
-	}
-
-	/**
-	 * A filter to apply authentication transformation as supplied by the
-	 * {@link LocalAuthenticationProvider}. Relies on the session ID being
-	 * updated to ensure that stale login information is not retained.
-	 *
-	 * @see LocalAuthenticationProvider#updateAuthentication(SecurityContext)
-	 */
-	@Component
-	public static class AppAuthTransformationFilter
-			extends OncePerRequestFilter {
-		/**
-		 * The name of session fixation tokens supported by this class.
-		 */
-		private static final String TOKEN =
-				AppAuthTransformationFilter.class.getName();
-
-		@Autowired
-		private LocalAuthenticationProvider localAuthProvider;
-
-		@Override
-		protected boolean shouldNotFilter(HttpServletRequest request)
-				throws ServletException {
-			Authentication a =
-					SecurityContextHolder.getContext().getAuthentication();
-			return isNull(a) || isUnsupportedAuthTokenClass(a.getClass());
-		}
-
-		@Override
-		protected void doFilterInternal(HttpServletRequest request,
-				HttpServletResponse response, FilterChain chain)
-				throws IOException, ServletException {
-			HttpSession s = request.getSession(false);
-			SecurityContext ctx = SecurityContextHolder.getContext();
-			Authentication savedAuth = getSavedToken(s);
-			Authentication originalAuth = ctx.getAuthentication();
-			if (nonNull(savedAuth)) {
-				ctx.setAuthentication(savedAuth);
-			} else {
-				Authentication a = localAuthProvider.updateAuthentication(ctx);
-				if (nonNull(a)) {
-					saveToken(s, a);
-				}
-			}
-			try {
-				chain.doFilter(request, response);
-			} finally {
-				ctx.setAuthentication(originalAuth);
-			}
-		}
-
-		private static Authentication getSavedToken(HttpSession session) {
-			if (nonNull(session)) {
-				Object o = session.getAttribute(TOKEN);
-				if (o instanceof Token) {
-					Token t = (Token) o;
-					if (t.isValid(session)) {
-						return t.getAuth();
-					}
-				}
-			}
-			return null;
-		}
-
-		private static void saveToken(HttpSession session,
-				Authentication auth) {
-			if (nonNull(session)) {
-				session.setAttribute(TOKEN, new Token(session, auth));
-			}
-		}
-
-		/**
-		 * Ensure that any authentication token that was bound into the session
-		 * is removed.
-		 *
-		 * @param request
-		 *            the HTTP request; used for looking up the session
-		 */
-		public static void clearToken(HttpServletRequest request) {
-			HttpSession session = request.getSession(false);
-			if (nonNull(session)) {
-				session.removeAttribute(TOKEN);
-			}
-		}
 	}
 
 	/**
